# docs/swagger.yaml
openapi: 3.0.2
info:
    title: books-api
    version: 1.0.0
    description: >-
        ### This is a Swagger UI for our Games API  

        ### Error handling

        Errors are presented using the HTTP status codes 400-599. The response includes a more detailed
        explanation about the error in the body.

        ```
        {
          "error": "Cleartext explanation of the error that occurred"
        }
        ```

        ### Generic API errors

        These error responses can be returned by any of the described endpoints.

        | HTTP status  |  Description |
        |---|---|
        | 400| Bad input parameters or invalid request. More information can be found in the response body.
        | 401| The consumer does not have access to the requested resource or the provided token is not valid.
        | 404| The requested resource cannot be found or the user does not have access to the resource.
        | 5XX| A server error has occurred while processing the request.
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of games:
                                    value: |-
                                        [
                                            "Witcher 3"
                                            "Cyberpunk 2077"
                                        ]
                    description: 'Returns a list of games.'
            summary: Get a list of all games
            description: Returns a list of all games in the API

    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                    name:
                                        type: string
                                    price:
                                        type: number
                            examples:
                                Full details of a given game:
                                    value: {"id": 1, "name": "Witcher 3", "price": 29.99}
                    description: OK
                '404':
                    description: Not found
            deprecated: false
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given game by its ID
        parameters:
            -
                name: id
                in: path
                required: true
                schema:
                    type: integer
        delete: # Add the DELETE operation here
            tags:
                - details
            responses:
                '204':
                    description: No Content
                '404':
                    description: Not found
            operationId: deleteGame
            summary: Delete a game
            description: Deletes a game by its ID
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    schema:
                        type: integer
components:
    schemas:
        error:
            title: Root Type for error
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.